refactor(prompt-detail): split PromptDetailView.tsx into focused modules (#94)
Extract 1,146-line monolith into 12 focused modules under
src/components/dashboard/prompt-detail/:
- types.ts: shared type definitions (EvidenceStatus, InjectedEvidenceItem)
- constants.ts: constants, color maps, utility functions, lazy imports
- evidence.ts: evidence classification logic (engine + legacy fallback)
- usePromptDetail.ts: hook for evidence loading and compaction counting
- StatPill.tsx, Section.tsx: reusable UI primitives
- SignalBreakdown.tsx, EvidenceGroup.tsx: evidence visualization components
- ActionFilterChips.tsx: tool filter chip bar
- FilePreviewOverlay.tsx: file preview modal with syntax highlighting
- ContextGauge.tsx: circular context usage gauge
- JourneySummary.tsx: prompt journey card grid

Result: PromptDetailView.tsx reduced from 1,146 to 228 lines.
